style: 잘못된 패키지 구조 수정
- spinner 클래스들 domain -> ui 로 이동
